perf: 카테고리별 책 통계 API 쿼리 변

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -145,10 +145,10 @@ class BookServiceTest @Autowired constructor (
         // then
         assertThat(results).hasSize(2)
         assertCount(results, BookType.COMPUTER, 2)
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.SCIENCE, 1L)
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
